Add text. Fix libs (again)
Added some text to get us started on this one as per discussion with Layla. We can extend it from here.

Also, the libs were broken again somehow, so I ended up needing to move it back to %pip install in the config notebook (except pytest). Hopefully this will be stable.